Bed cleaning system implementation, paying manager improvement
Player can not retrieve sheets and pajamas from used beds to be dropped off in the laundry.

Laundry is also now functional.

Customer movement to the paying line has improved that even though the line moves up while he/she is in transit, he/she will not stay in the node he originally intended to go to.