Penalty system, paying scheme, suitcase return scheme implementation
Starting development of the payment scheme by creating the satisfaction index system and by introducing a penalty system (ticket-issuing system).

Fixed the return of objects so that dirty pajamas cannot be given to new customers, etc.

Set up the patience and alarm systems to be connected to the customer satisfaction. Also fixed the resetting of timers, computation of merits and demerits.