UI modification
Decided how the UI will change whenever a bed is occupied, for example.. or when a player receives pajamas or when the luggage is given, etc.

Will be changing the sprites instead of using tilemaps and specifying transforms then changing the tile on the transform's location. Will be using the scriptable objects with this for convenience.